+++ rs/pocket_ic_server/src/state_api/state.rs
@@ -289,6 +289,7 @@ pub enum PocketIcError {
     TooManyMockCanisterHttpResponses((usize, usize)),
     InvalidRejectCode(u64),
     CanisterHttpRejectMessageTooLong((usize, usize)),
+    CanisterHttpSpentCyclesTooLarge((u128, u128)),
     SettingTimeIntoPast((u64, u64)),
     Forbidden(String),
     BlockmakerNotFound(NodeId),
@@ -373,6 +374,12 @@ impl std::fmt::Debug for OpOut {
                 write!(
                     f,
                     "CanisterHttpRejectMessageTooLong(actual={actual},max={max})"
+                )
+            }
+            OpOut::Error(PocketIcError::CanisterHttpSpentCyclesTooLarge((actual, max))) => {
+                write!(
+                    f,
+                    "CanisterHttpSpentCyclesTooLarge(actual={actual},max={max})"
                 )
             }
             OpOut::Error(PocketIcError::SettingTimeIntoPast((current, set))) => {